feat: proof of concept for song chances

+++ include/overworld.h
@@ -127,7 +127,7 @@ u16 GetCurrLocationDefaultMusic(void);
 u16 GetWarpDestinationMusic(void);
 void Overworld_ResetMapMusic(void);
 void Overworld_PlaySpecialMapMusic(void);
-void Overworld_SetSavedMusic(u16 songNum);
+void Overworld_SetSavedMusic(u32 songNum);
 void Overworld_ClearSavedMusic(void);
 void Overworld_ChangeMusicToDefault(void);
 void Overworld_ChangeMusicTo(u16 newMusic);
